+++ controllers/openshift_controller.go
@@ -20,7 +20,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"os"
-	"path/filepath"
 	"reflect"
 	"strings"
 	"time"
@@ -28,7 +27,6 @@ import (
 	appsv1 "k8s.io/api/apps/v1"
 
 	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/apimachinery/pkg/util/intstr"
 
 	ignTypes "github.com/coreos/ignition/v2/config/v3_2/types"
 	"github.com/go-logr/logr"
@@ -68,30 +66,17 @@ type KataConfigOpenShiftReconciler struct {
 }
 
 const (
-	OperatorNamespace                   = "openshift-sandboxed-containers-operator"
-	dashboard_configmap_name            = "grafana-dashboard-sandboxed-containers"
-	dashboard_configmap_namespace       = "openshift-config-managed"
-	container_runtime_config_name       = "kata-crio-config"
-	extension_mc_name                   = "50-enable-sandboxed-containers-extension"
-	DEFAULT_PEER_PODS                   = "10"
-	peerpodConfigCrdName                = "peerpodconfig-openshift"
-	peerpodsMachineConfigPathLocation   = "/config/peerpods"
-	peerpodsCrioMachineConfig           = "50-kata-remote"
-	peerpodsCrioMachineConfigYaml       = "mc-50-crio-config.yaml"
-	peerpodsKataRemoteMachineConfig     = "40-worker-kata-remote-config"
-	peerpodsKataRemoteMachineConfigYaml = "mc-40-kata-remote-config.yaml"
+	OperatorNamespace             = "openshift-sandboxed-containers-operator"
+	dashboard_configmap_name      = "grafana-dashboard-sandboxed-containers"
+	dashboard_configmap_namespace = "openshift-config-managed"
+	container_runtime_config_name = "kata-crio-config"
+	extension_mc_name             = "50-enable-sandboxed-containers-extension"
 	// Use same Pod Overhead as upstream kata-deploy using, see
 	// https://github.com/kata-containers/kata-containers/blob/main/tools/packaging/kata-deploy/runtimeclasses/kata-qemu.yaml#L7
 	kataRuntimeClassName        = "kata"
 	kataRuntimeClassCpuOverhead = "0.25"
 	// We need a higher value than upstream (see https://github.com/openshift/sandboxed-containers-operator/pull/84)
 	kataRuntimeClassMemOverhead = "350Mi"
-	// https://github.com/kata-containers/kata-containers/blob/main/tools/packaging/kata-deploy/runtimeclasses/kata-remote.yaml#L7
-	peerpodsRuntimeClassName        = "kata-remote"
-	peerpodsRuntimeClassCpuOverhead = "0.25"
-	peerpodsRuntimeClassMemOverhead = "120Mi"
-	// cloud-api-adaptor (CAA) daemonset name
-	caaDsName = "osc-caa-ds"
 )
 
 // +kubebuilder:rbac:groups=kataconfiguration.openshift.io,resources=kataconfigs;kataconfigs/finalizers,verbs=get;list;watch;create;update;patch;delete
@@ -331,261 +316,6 @@ func (r *KataConfigOpenShiftReconciler) removeLogLevel() error {
 	}
 	r.Log.Info("logLevel ContainerRuntimeConfig deleted successfully")
 	return nil
-}
-
-func MountProgagationRef(mode corev1.MountPropagationMode) *corev1.MountPropagationMode {
-	return &mode
-}
-
-func (r *KataConfigOpenShiftReconciler) processDaemonsetForCAA() *appsv1.DaemonSet {
-	var (
-		runPrivileged                = true
-		runAsUser              int64 = 0
-		defaultMode            int32 = 0600
-		sshSecretOptional            = true
-		authJsonSecretOptional       = true
-		nodeSelector                 = r.getNodeSelectorAsMap()
-	)
-
-	dsLabelSelectors := map[string]string{
-		"name": caaDsName,
-	}
-
-	imageString := os.Getenv("RELATED_IMAGE_CAA")
-	if imageString == "" {
-		r.Log.Info("RELATED_IMAGE_CAA env var is unset or empty, cloud-api-adaptor pods will not run")
-	}
-
-	return &appsv1.DaemonSet{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: "apps/v1",
-			Kind:       "DaemonSet",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      caaDsName,
-			Namespace: os.Getenv("PEERPODS_NAMESPACE"),
-		},
-		Spec: appsv1.DaemonSetSpec{
-			Selector: &metav1.LabelSelector{
-				MatchLabels: dsLabelSelectors,
-			},
-			UpdateStrategy: appsv1.DaemonSetUpdateStrategy{
-				Type: "RollingUpdate",
-				RollingUpdate: &appsv1.RollingUpdateDaemonSet{
-					MaxUnavailable: &intstr.IntOrString{
-						Type:   intstr.Int,
-						IntVal: 1,
-					},
-				},
-			},
-			Template: corev1.PodTemplateSpec{
-				ObjectMeta: metav1.ObjectMeta{
-					Labels: dsLabelSelectors,
-				},
-				Spec: corev1.PodSpec{
-					ServiceAccountName: "default",
-					NodeSelector:       nodeSelector,
-					HostNetwork:        true,
-					Containers: []corev1.Container{
-						{
-							Name:            "caa-pod",
-							Image:           imageString,
-							ImagePullPolicy: "Always",
-							SecurityContext: &corev1.SecurityContext{
-								// TODO - do we really need to run as root?
-								Privileged: &runPrivileged,
-								RunAsUser:  &runAsUser,
-							},
-							Command: []string{"/usr/local/bin/entrypoint.sh"},
-							Env: []corev1.EnvVar{
-								{
-									Name: "NODE_NAME",
-									ValueFrom: &corev1.EnvVarSource{
-										FieldRef: &corev1.ObjectFieldSelector{
-											FieldPath: "spec.nodeName",
-										},
-									},
-								},
-							},
-							EnvFrom: []corev1.EnvFromSource{
-								{
-									SecretRef: &corev1.SecretEnvSource{
-										LocalObjectReference: corev1.LocalObjectReference{
-											Name: "peer-pods-secret",
-										},
-									},
-								},
-								{
-									ConfigMapRef: &corev1.ConfigMapEnvSource{
-										LocalObjectReference: corev1.LocalObjectReference{
-											Name: "peer-pods-cm",
-										},
-									},
-								},
-							},
-							VolumeMounts: []corev1.VolumeMount{
-								{
-									Name:      "auth-json-volume",
-									MountPath: "/root/containers/",
-									ReadOnly:  true,
-								}, {
-									Name:      "ssh",
-									MountPath: "/root/.ssh",
-									ReadOnly:  true,
-								},
-								{
-									MountPath: "/run/peerpod",
-									Name:      "pods-dir",
-								},
-								{
-									MountPath:        "/run/netns",
-									MountPropagation: MountProgagationRef(corev1.MountPropagationHostToContainer),
-									Name:             "netns",
-								},
-							},
-						},
-					},
-					Volumes: []corev1.Volume{
-						{
-							Name: "auth-json-volume",
-							VolumeSource: corev1.VolumeSource{
-								Secret: &corev1.SecretVolumeSource{
-									SecretName:  "auth-json-secret",
-									DefaultMode: &defaultMode,
-									Optional:    &authJsonSecretOptional,
-								},
-							},
-						}, {
-							Name: "ssh",
-							VolumeSource: corev1.VolumeSource{
-								Secret: &corev1.SecretVolumeSource{
-									SecretName:  "ssh-key-secret",
-									DefaultMode: &defaultMode,
-									Optional:    &sshSecretOptional,
-								},
-							},
-						},
-						{
-							Name: "pods-dir",
-							VolumeSource: corev1.VolumeSource{
-								HostPath: &corev1.HostPathVolumeSource{
-									Path: "/run/peerpod",
-								},
-							},
-						},
-						{
-							Name: "netns",
-							VolumeSource: corev1.VolumeSource{
-								HostPath: &corev1.HostPathVolumeSource{
-									Path: "/run/netns",
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-}
-
-func (r *KataConfigOpenShiftReconciler) processDaemonsetForMonitor() *appsv1.DaemonSet {
-	var (
-		runPrivileged = false
-		runUserID     = int64(1001)
-		runGroupID    = int64(1001)
-	)
-
-	kataMonitorImage := os.Getenv("RELATED_IMAGE_KATA_MONITOR")
-	if len(kataMonitorImage) == 0 {
-		// kata-monitor image URL is generally impossible to verify or sanitise,
-		// with the empty value being pretty much the only exception where it's
-		// fairly clear what good it is.  If we can only detect a single one
-		// out of an infinite number of bad values, we choose not to return an
-		// error here (giving an impression that we can actually detect errors)
-		// but just log this incident and plow ahead.
-		r.Log.Info("RELATED_IMAGE_KATA_MONITOR env var is unset or empty, kata-monitor pods will not run")
-	}
-
-	r.Log.Info("Creating monitor DaemonSet with image file: \"" + kataMonitorImage + "\"")
-	dsName := "openshift-sandboxed-containers-monitor"
-	dsLabels := map[string]string{
-		"name": dsName,
-	}
-
-	nodeSelector := r.getNodeSelectorAsMap()
-
-	return &appsv1.DaemonSet{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: "apps/v1",
-			Kind:       "DaemonSet",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      dsName,
-			Namespace: OperatorNamespace,
-		},
-		Spec: appsv1.DaemonSetSpec{
-			Selector: &metav1.LabelSelector{
-				MatchLabels: dsLabels,
-			},
-			Template: corev1.PodTemplateSpec{
-				ObjectMeta: metav1.ObjectMeta{
-					Labels: dsLabels,
-				},
-				Spec: corev1.PodSpec{
-					ServiceAccountName: "monitor",
-					NodeSelector:       nodeSelector,
-					Tolerations: []corev1.Toleration{
-						{
-							Operator: corev1.TolerationOpExists,
-						},
-					},
-					Containers: []corev1.Container{
-						{
-							Name:            "kata-monitor",
-							Image:           kataMonitorImage,
-							ImagePullPolicy: "Always",
-							SecurityContext: &corev1.SecurityContext{
-								Privileged: &runPrivileged,
-								RunAsUser:  &runUserID,
-								RunAsGroup: &runGroupID,
-								SELinuxOptions: &corev1.SELinuxOptions{
-									Type: "osc_monitor.process",
-								},
-							},
-							Command: []string{"/usr/bin/kata-monitor", "--listen-address=:8090", "--log-level=debug", "--runtime-endpoint=/run/crio/crio.sock"},
-							VolumeMounts: []corev1.VolumeMount{
-								{
-									Name:      "crio-sock",
-									MountPath: "/run/crio/",
-								},
-								{
-									Name:      "sbs",
-									MountPath: "/run/vc/sbs/",
-								}},
-						},
-					},
-					Volumes: []corev1.Volume{
-						{
-							Name: "crio-sock",
-							VolumeSource: corev1.VolumeSource{
-								HostPath: &corev1.HostPathVolumeSource{
-									Path: "/run/crio/",
-								},
-							},
-						},
-						{
-							Name: "sbs",
-							VolumeSource: corev1.VolumeSource{
-								HostPath: &corev1.HostPathVolumeSource{
-									Path: "/run/vc/sbs/",
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-	}
 }
 
 func (r *KataConfigOpenShiftReconciler) processDashboardConfigMap() *corev1.ConfigMap {
@@ -2365,186 +2095,6 @@ func (r *KataConfigOpenShiftReconciler) isUpdating() bool {
 		return false
 	}
 	return cond.Status == corev1.ConditionTrue && cond.Reason == "Updating"
-}
-
-func (r *KataConfigOpenShiftReconciler) createAuthJsonSecret() error {
-	var err error
-
-	pullSecret := &corev1.Secret{}
-	err = r.Client.Get(context.TODO(), types.NamespacedName{Name: "pull-secret", Namespace: "openshift-config"}, pullSecret)
-	if err != nil {
-		r.Log.Info("Error fetching pull-secret", "err", err)
-		return err
-	}
-
-	authJsonSecret := corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "auth-json-secret",
-			Namespace: OperatorNamespace,
-		},
-		Data: map[string][]byte{
-			"auth.json": pullSecret.Data[".dockerconfigjson"],
-		},
-		Type: corev1.SecretTypeOpaque,
-	}
-
-	err = r.Client.Create(context.TODO(), &authJsonSecret)
-	if err != nil {
-		if k8serrors.IsAlreadyExists(err) {
-			err = r.Client.Update(context.TODO(), &authJsonSecret)
-			if err != nil {
-				r.Log.Info("Error updating auth-json-secret", "err", err)
-				return err
-			}
-		} else {
-			r.Log.Info("Error creating auth-json-secret", "err", err)
-			return err
-		}
-	}
-
-	return err
-}
-
-// Create the MachineConfigs for PeerPod
-// We do it before kata-oc creation to optimise the reboots required for MC creation
-func (r *KataConfigOpenShiftReconciler) enablePeerPodsMc() error {
-
-	//Create MachineConfig for kata-remote hyp CRIO config
-	crioMachineConfigFilePath := filepath.Join(peerpodsMachineConfigPathLocation, peerpodsCrioMachineConfigYaml)
-	err := r.createMcFromFile(crioMachineConfigFilePath)
-	if err != nil {
-		r.Log.Info("Error in creating CRIO MachineConfig", "err", err)
-		return err
-	}
-
-	//Create MachineConfig for kata-remote hyp config toml
-	kataConfigMachineConfigFilePath := filepath.Join(peerpodsMachineConfigPathLocation, peerpodsKataRemoteMachineConfigYaml)
-	err = r.createMcFromFile(kataConfigMachineConfigFilePath)
-	if err != nil {
-		r.Log.Info("Error in creating kata remote configuration.toml MachineConfig", "err", err)
-		return err
-	}
-
-	return nil
-}
-
-// Create the PeerPodConfig CRDs and misc configs required for peer-pods
-func (r *KataConfigOpenShiftReconciler) enablePeerPodsMiscConfigs() error {
-	// Create the CAA daemonset
-	ds := r.processDaemonsetForCAA()
-	if err := controllerutil.SetControllerReference(r.kataConfig, ds, r.Scheme); err != nil {
-		r.Log.Error(err, "Failed setting ControllerReference for cloud-api-adaptor DS")
-		return err
-	}
-
-	err := r.Client.Update(context.TODO(), ds)
-	if err != nil && k8serrors.IsNotFound(err) {
-		r.Log.Error(err, "cloud-api-adaptor daemonset doesn't exist. Creating")
-		err = r.Client.Create(context.TODO(), ds)
-		if err != nil {
-			r.Log.Error(err, "failed to create cloud-api-adaptor daemonset")
-			return err
-		}
-	}
-
-	// Create the mutating webhook deployment
-	err = r.createMutatingWebhookDeployment()
-	if err != nil {
-		r.Log.Info("Error in creating mutating webhook deployment for peerpods", "err", err)
-		return err
-	}
-
-	// Create the mutating webhook service
-	err = r.createMutatingWebhookService()
-	if err != nil {
-		r.Log.Info("Error in creating mutating webhook service for peerpods", "err", err)
-		return err
-	}
-
-	// Create the mutating webhook
-	err = r.createMutatingWebhookConfig()
-	if err != nil {
-		r.Log.Info("Error in creating mutating webhook for peerpods", "err", err)
-		return err
-	}
-
-	// Create runtimeClass config for peer-pods
-	err = r.createRuntimeClass(peerpodsRuntimeClassName, peerpodsRuntimeClassCpuOverhead, peerpodsRuntimeClassMemOverhead)
-	if err != nil {
-		r.Log.Info("Error in creating kata remote runtimeclass", "err", err)
-		return err
-	}
-	return nil
-}
-
-func (r *KataConfigOpenShiftReconciler) disablePeerPods() error {
-	ds := r.processDaemonsetForCAA()
-	err := r.Client.Delete(context.TODO(), ds)
-	if err != nil {
-		if k8serrors.IsNotFound(err) {
-			r.Log.Info("cloud-api-adaptor daemonset was already deleted")
-		} else {
-			r.Log.Error(err, "error when deleting cloud-api-adaptor Daemonset, try again")
-			return err
-		}
-	}
-
-	mc := mcfgv1.MachineConfig{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: "machineconfiguration.openshift.io/v1",
-			Kind:       "MachineConfig",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name: peerpodsKataRemoteMachineConfig,
-		},
-	}
-
-	err = r.Client.Delete(context.TODO(), &mc)
-	if err != nil {
-		// error during removing mc. Just log the error and move on.
-		r.Log.Info("Error found deleting mc. If the MachineConfig object exists after uninstallation it can be safely deleted manually",
-			"mc", mc.Name, "err", err)
-	}
-
-	mc = mcfgv1.MachineConfig{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: "machineconfiguration.openshift.io/v1",
-			Kind:       "MachineConfig",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name: peerpodsCrioMachineConfig,
-		},
-	}
-
-	err = r.Client.Delete(context.TODO(), &mc)
-	if err != nil {
-		// error during removing mc. Just log the error and move on.
-		r.Log.Info("Error found deleting mc. If the MachineConfig object exists after uninstallation it can be safely deleted manually",
-			"mc", mc.Name, "err", err)
-	}
-
-	// Delete mutating webhook deployment
-	err = r.deleteMutatingWebhookDeployment()
-	if err != nil {
-		r.Log.Info("Error in deleting mutating webhook deployment for peerpods", "err", err)
-		return err
-	}
-
-	// Delete mutating webhook service
-	err = r.deleteMutatingWebhookService()
-	if err != nil {
-		r.Log.Info("Error in deleting mutating webhook service for peerpods", "err", err)
-		return err
-	}
-
-	// Delete the mutating webhook
-	err = r.deleteMutatingWebhookConfig()
-	if err != nil {
-		r.Log.Info("Error in deleting mutating webhook for peerpods", "err", err)
-		return err
-	}
-
-	return nil
 }
 
 // Create the MachineConfigs from file
